agent/requirements.txt
# LiveKit Agent dependencies
livekit
livekit-agents
livekit-plugins-openai
python-dotenv
requests
